- use const on more occasions (e.g. result of Ns_ConfigGetPath) - reduce implicit signed/unsigned conversions - use boolean expressions in conditions with flags - adding nonnull assertions (connio) - avoid reuse of C90 identifier pattern (str[a-z]+)

+++ nsd/limits.c
@@ -233,10 +233,10 @@ NsTclSetLimitsObjCmd(ClientData UNUSED(clientData), Tcl_Interp *interp, int objc
         return TCL_ERROR;
     }
     if (maxrun > -1) {
-        limitsPtr->maxrun = maxrun;
+	limitsPtr->maxrun = (unsigned int)maxrun;
     }
     if (maxwait > -1) {
-        limitsPtr->maxwait = maxwait;
+        limitsPtr->maxwait = (unsigned int)maxwait;
     }
     if (maxupload > -1) {
 	limitsPtr->maxupload = (size_t)maxupload;
@@ -333,8 +333,8 @@ FindLimits(const char *limits, int create)
             limitsPtr->name = Tcl_GetHashKey(&limtable, hPtr);
             Ns_MutexInit(&limitsPtr->lock);
             Ns_MutexSetName2(&limitsPtr->lock, "ns:limits", limits);
-            limitsPtr->maxrun = limitsPtr->maxwait = 100;
-            limitsPtr->maxupload = 10 * 1024 * 1000; /* NB: 10meg limit. */
+            limitsPtr->maxrun = limitsPtr->maxwait = 100U;
+            limitsPtr->maxupload = 10U * 1024U * 1000U; /* NB: 10meg limit. */
             limitsPtr->timeout = 60;
             Tcl_SetHashValue(hPtr, limitsPtr);
         }
